Folder reorganization and test of the inputs and bug corrections
