+++ Unity.RenderPipelines.Core.Editor.csproj
@@ -871,15 +871,12 @@
     <Reference Include="System.Xml.XPath.XDocument">
       <HintPath>D:\Unity\2021.3.13f1\Editor\Data\UnityReferenceAssemblies\unity-4.8-api\Facades\System.Xml.XPath.XDocument.dll</HintPath>
     </Reference>
-<<<<<<< HEAD
     <Reference Include="UnityEditor.TestRunner">
       <HintPath>Library\ScriptAssemblies\UnityEditor.TestRunner.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine.TestRunner">
       <HintPath>Library\ScriptAssemblies\UnityEngine.TestRunner.dll</HintPath>
     </Reference>
-=======
->>>>>>> origin/develop
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Unity.RenderPipelines.Core.Runtime.csproj">
@@ -894,17 +891,6 @@
       <Project>{C11F3F69-389C-75CE-0FF8-88B73F5E3791}</Project>
       <Name>UnityEngine.UI</Name>
     </ProjectReference>
-<<<<<<< HEAD
-=======
-    <ProjectReference Include="UnityEditor.TestRunner.csproj">
-      <Project>{F8222DA0-E985-4803-B661-D2A52BEEFAA1}</Project>
-      <Name>UnityEditor.TestRunner</Name>
-    </ProjectReference>
-    <ProjectReference Include="UnityEngine.TestRunner.csproj">
-      <Project>{8A4D20E2-E2CB-407F-B76B-34015CF1962E}</Project>
-      <Name>UnityEngine.TestRunner</Name>
-    </ProjectReference>
->>>>>>> origin/develop
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="GenerateTargetFrameworkMonikerAttribute" />
