+++ Unity.RenderPipelines.Core.Runtime.Player.csproj
@@ -806,20 +806,11 @@
     <Reference Include="System.Xml.XPath.XDocument">
       <HintPath>D:\Unity\2021.3.13f1\Editor\Data\UnityReferenceAssemblies\unity-4.8-api\Facades\System.Xml.XPath.XDocument.dll</HintPath>
     </Reference>
-<<<<<<< HEAD
     <Reference Include="Unity.InputSystem">
       <HintPath>Library\ScriptAssemblies\Unity.InputSystem.dll</HintPath>
     </Reference>
   </ItemGroup>
   <ItemGroup>
-=======
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="Unity.InputSystem.Player.csproj">
-      <Project>{FC1BA666-6609-0CED-6D13-091F66C6D008}</Project>
-      <Name>Unity.InputSystem.Player</Name>
-    </ProjectReference>
->>>>>>> origin/develop
     <ProjectReference Include="UnityEngine.UI.Player.csproj">
       <Project>{AC4EC7D3-B2DC-0933-D334-2BFD1B99BA98}</Project>
       <Name>UnityEngine.UI.Player</Name>
