+++ Unity.RenderPipelines.Universal.Runtime.Player.csproj
@@ -824,15 +824,12 @@
     <Reference Include="System.Xml.XPath.XDocument">
       <HintPath>D:\Unity\2021.3.13f1\Editor\Data\UnityReferenceAssemblies\unity-4.8-api\Facades\System.Xml.XPath.XDocument.dll</HintPath>
     </Reference>
-<<<<<<< HEAD
     <Reference Include="Unity.Mathematics">
       <HintPath>Library\ScriptAssemblies\Unity.Mathematics.dll</HintPath>
     </Reference>
     <Reference Include="Unity.Burst">
       <HintPath>Library\ScriptAssemblies\Unity.Burst.dll</HintPath>
     </Reference>
-=======
->>>>>>> origin/develop
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Unity.RenderPipelines.Core.Runtime.Player.csproj">
@@ -843,17 +840,6 @@
       <Project>{C88631E1-6DCB-39F7-BC05-ABCCB942AFBD}</Project>
       <Name>Unity.RenderPipeline.Universal.ShaderLibrary.Player</Name>
     </ProjectReference>
-<<<<<<< HEAD
-=======
-    <ProjectReference Include="Unity.Mathematics.Player.csproj">
-      <Project>{0D6CACB8-8939-31EE-5317-0E28305E8345}</Project>
-      <Name>Unity.Mathematics.Player</Name>
-    </ProjectReference>
-    <ProjectReference Include="Unity.Burst.Player.csproj">
-      <Project>{43B42743-C800-ECF4-1CB0-034116511CF4}</Project>
-      <Name>Unity.Burst.Player</Name>
-    </ProjectReference>
->>>>>>> origin/develop
     <ProjectReference Include="UnityEngine.UI.Player.csproj">
       <Project>{AC4EC7D3-B2DC-0933-D334-2BFD1B99BA98}</Project>
       <Name>UnityEngine.UI.Player</Name>
