+++ Unity.RenderPipelines.Universal.Runtime.csproj
@@ -899,15 +899,12 @@
     <Reference Include="System.Xml.XPath.XDocument">
       <HintPath>D:\Unity\2021.3.13f1\Editor\Data\UnityReferenceAssemblies\unity-4.8-api\Facades\System.Xml.XPath.XDocument.dll</HintPath>
     </Reference>
-<<<<<<< HEAD
     <Reference Include="Unity.Mathematics">
       <HintPath>Library\ScriptAssemblies\Unity.Mathematics.dll</HintPath>
     </Reference>
     <Reference Include="Unity.Burst">
       <HintPath>Library\ScriptAssemblies\Unity.Burst.dll</HintPath>
     </Reference>
-=======
->>>>>>> origin/develop
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Unity.RenderPipelines.Core.Runtime.csproj">
@@ -918,17 +915,6 @@
       <Project>{32CA72BF-E503-A6FE-4BB0-61A20EAD13C1}</Project>
       <Name>Unity.RenderPipeline.Universal.ShaderLibrary</Name>
     </ProjectReference>
-<<<<<<< HEAD
-=======
-    <ProjectReference Include="Unity.Mathematics.csproj">
-      <Project>{70CB6951-E018-6113-089B-3FF33C21761B}</Project>
-      <Name>Unity.Mathematics</Name>
-    </ProjectReference>
-    <ProjectReference Include="Unity.Burst.csproj">
-      <Project>{021067B4-92BA-2D27-9A9F-FBCA71B66476}</Project>
-      <Name>Unity.Burst</Name>
-    </ProjectReference>
->>>>>>> origin/develop
     <ProjectReference Include="UnityEditor.UI.csproj">
       <Project>{C78F8C90-B95D-3908-556F-8CC255FD5B7A}</Project>
       <Name>UnityEditor.UI</Name>
