+++ Unity.ShaderGraph.Editor.csproj
@@ -1497,7 +1497,6 @@
     <Reference Include="System.Xml.XPath.XDocument">
       <HintPath>D:\Unity\2021.3.13f1\Editor\Data\UnityReferenceAssemblies\unity-4.8-api\Facades\System.Xml.XPath.XDocument.dll</HintPath>
     </Reference>
-<<<<<<< HEAD
     <Reference Include="Unity.Searcher.Editor">
       <HintPath>Library\ScriptAssemblies\Unity.Searcher.Editor.dll</HintPath>
     </Reference>
@@ -1507,21 +1506,12 @@
     <Reference Include="UnityEngine.TestRunner">
       <HintPath>Library\ScriptAssemblies\UnityEngine.TestRunner.dll</HintPath>
     </Reference>
-=======
->>>>>>> origin/develop
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="Unity.ShaderGraph.Utilities.csproj">
       <Project>{6C84049C-AA35-AEEA-0091-8E66A15EA67F}</Project>
       <Name>Unity.ShaderGraph.Utilities</Name>
     </ProjectReference>
-<<<<<<< HEAD
-=======
-    <ProjectReference Include="Unity.Searcher.Editor.csproj">
-      <Project>{8DAEACA9-2E8D-A244-4DF6-DBA6048D89AD}</Project>
-      <Name>Unity.Searcher.Editor</Name>
-    </ProjectReference>
->>>>>>> origin/develop
     <ProjectReference Include="Unity.RenderPipelines.Core.Runtime.csproj">
       <Project>{7548AA12-C1E4-754E-A47D-4A5D7D2F1E07}</Project>
       <Name>Unity.RenderPipelines.Core.Runtime</Name>
@@ -1538,17 +1528,6 @@
       <Project>{C11F3F69-389C-75CE-0FF8-88B73F5E3791}</Project>
       <Name>UnityEngine.UI</Name>
     </ProjectReference>
-<<<<<<< HEAD
-=======
-    <ProjectReference Include="UnityEditor.TestRunner.csproj">
-      <Project>{F8222DA0-E985-4803-B661-D2A52BEEFAA1}</Project>
-      <Name>UnityEditor.TestRunner</Name>
-    </ProjectReference>
-    <ProjectReference Include="UnityEngine.TestRunner.csproj">
-      <Project>{8A4D20E2-E2CB-407F-B76B-34015CF1962E}</Project>
-      <Name>UnityEngine.TestRunner</Name>
-    </ProjectReference>
->>>>>>> origin/develop
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="GenerateTargetFrameworkMonikerAttribute" />
